# openapi.yaml
...
openapi: 3.1.0
info:
  title: Silicon Hook API
  version: 0.5.0
  summary: Signed webhook endpoints with ordered, acknowledged delivery to Silicons.
  description: |
    Silicon Hook gives each Silicon signed webhook endpoints at
    `https://hook.teamofsilicons.com/silicon/{silicon_id}/{endpoint_key}`.
    Every request to an active endpoint is answered with `webhook.ok`. Requests
    that verify against the hook's signature policy are retained for 14 days
    and delivered in order over the WebSocket stream at `/api/v1/ws` or by
    polling `/api/v1/silicons/{silicon_id}/deliveries`; requests that do not
    verify are withheld and kept in a separate 14-day blocked log. Timestamps
    are UTC RFC 3339 values. Endpoint keys route requests but never
    authenticate them. IAM test app_secret selectors choose isolated sandboxes via
    X-Hook-Test-App-Secret; they never grant actor authority. Deprecated API v1
    contracts sunset only after seven request-free days. Contract accounting is
    independent of optional diagnostic telemetry.
  contact:
    email: engineering@teamofsilicons.com
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
- url: https://backend.hook.teamofsilicons.com
  description: Backend API; public ingress may use the configured Hook frontend origin.
security:
- bearerAuth: []
...
paths:
...
  "/api/v1/silicons/{silicon_id}/hooks":
    get:
      tags:
      - Hooks
      operationId: listHooks
      summary: List a Silicon's hooks with their URLs and last activity
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - "$ref": "#/components/parameters/SiliconId"
      - name: include_deleted
        in: query
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: Active and disabled hooks visible to the actor, plus recoverable
            deleted hooks when include_deleted is true. Each item carries its name,
            endpoint URL, and last_received_at.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HookPage"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    post:
      tags:
      - Hooks
      operationId: createHook
      summary: Create a signed webhook endpoint
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - "$ref": "#/components/parameters/SiliconId"
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/HookCreate"
      responses:
        '201':
          description: Hook created; the signing secret is shown once.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HookWithSecret"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '409':
          "$ref": "#/components/responses/Conflict"
        '410':
          "$ref": "#/components/responses/Gone"
        '413':
          "$ref": "#/components/responses/PayloadTooLarge"
        '415':
          "$ref": "#/components/responses/UnsupportedMediaType"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    patch:
      tags:
      - Hooks
      operationId: setHooksEnabled
      summary: Enable or disable a set of hooks atomically
      description: Applies the desired enabled state to 1–1000 unique retained hooks.
        The operation is atomic and intrinsically idempotent; returned items follow
        the request hook_ids order.
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - "$ref": "#/components/parameters/SiliconId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/HooksEnabledUpdate"
      responses:
        '200':
          description: All requested hooks now have the requested enabled state.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HookPage"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '409':
          "$ref": "#/components/responses/Conflict"
        '413':
          "$ref": "#/components/responses/PayloadTooLarge"
        '415':
          "$ref": "#/components/responses/UnsupportedMediaType"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    parameters:
    - "$ref": "#/components/parameters/TestKey"
    - "$ref": "#/components/parameters/TestAppSecret"
...
  "/api/v1/ws":
    get:
      tags:
      - Deliveries
      operationId: openDeliveryStream
      summary: Open the WebSocket delivery stream
      description: |-
        Authenticate the upgrade request exactly like a management call and name
        the Silicon streams with repeated `silicon_id` query parameters. After
        the `ready` frame the server replays every event after the consumer's
        acknowledged cursor, then streams live events. Frames are JSON text:

        - server `ready`, `ping`, `new_event`, `ack_recorded`, `error`
        - client `pong`, `ack`, `resume`

        The server pings every 30 seconds; two minutes without a matching pong
        closes the socket with code 4000 and reason `heartbeat-timeout`.

        Each hook event has exactly two top-level fields: `type: new_event` and
        `data`. The data object contains `sender` (the hook provider name at
        receipt) and `metadata` (the complete Event schema, including silicon_id,
        delivery_sequence, summary and captured request). This shape also applies
        to replays and client/CLI recipient POSTs. Control frames and REST Event
        objects retain their existing shapes.

        At most 32 outstanding events per Silicon. ACK releases capacity. Environment changes close with 4001 environment-changed; current IAM authorization failure closes with 4003 authorization-changed. Authorization is refreshed every 30 seconds and on cross-replica IAM webhook invalidation; a temporary IAM failure closes with 1013 iam-unavailable.
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - name: silicon_id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: array
          minItems: 1
          maxItems: 256
          items:
            "$ref": "#/components/schemas/SiliconId"
      responses:
        '101':
          description: Switching to the WebSocket delivery protocol.
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '426':
          description: The request is not a WebSocket upgrade.
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    parameters:
    - "$ref": "#/components/parameters/TestKey"
    - "$ref": "#/components/parameters/TestAppSecret"
...
  "/api/v1/testing-environment":
    get:
      tags:
      - Testing
      operationId: currentTestingEnvironment
      summary: Read root metadata using the Hook test key
      parameters:
      - &1
        name: X-Hook-Test-Key
        in: header
        required: true
        description: Selects an isolated Hook environment. The bearer must be from
          its linked IAM test world. Never attach to public ingress.
        schema:
          type: string
          pattern: "^[A-Za-z0-9]{32}$"
      responses:
        '200':
          description: Operation completed.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TestEnvironment"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      security: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: IAM opaque access token
  parameters:
    TestAppSecret:
      name: X-Hook-Test-App-Secret
      in: header
      required: false
      description: >-
        IAM test application selector, validated online on each request. Mutually
        exclusive with X-Hook-Test-Key. Never send on ingress or root administration
        routes. Invalid selectors fail closed; they never select production.
      schema:
        type: string
        pattern: '^ask_[A-Za-z0-9_-]{43}$'
    OrgId:
      name: X-Org-ID
      in: header
      required: true
      schema:
        "$ref": "#/components/schemas/OrganizationId"
    SiliconId:
      name: silicon_id
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/SiliconId"
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 255
        pattern: "^[!-~]+$"
...
    TestKey:
      name: X-Hook-Test-Key
      in: header
      required: false
      description: Selects an isolated Hook environment. The bearer must be from its
        linked IAM test world. Never attach to public ingress.
      schema:
        type: string
        pattern: "^[A-Za-z0-9]{32}$"
...
  responses:
    BadRequest:
      description: Required HTTP syntax, headers, or JSON are malformed.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
...
    Unauthorized:
      description: Credential is absent, invalid, expired, or revoked.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    Forbidden:
      description: The authenticated actor lacks authority.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
...
    NotFound:
      description: The resource does not exist or is not visible.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    Conflict:
      description: Idempotency or resource-state conflict.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    Gone:
      description: A recovery or one-time-secret replay window expired, or the endpoint
        key was retired by rotation (code endpoint_retired).
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    RequestTimeout:
      description: Request processing exceeded the service deadline.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    PayloadTooLarge:
      description: The body exceeds 1 MiB or the headers exceed 64 KiB.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    UnsupportedMediaType:
      description: The endpoint requires application/json.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    UnprocessableEntity:
      description: The request violates input rules; details may explain why.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    InternalError:
      description: An unexpected internal invariant or programming failure occurred.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    ServiceUnavailable:
      description: A required dependency cannot make a safe decision.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
  schemas:
...
    OrganizationId:
      type: string
      minLength: 1
      maxLength: 100
      pattern: "^[!-~]+$"
    SiliconId:
      type: string
      minLength: 1
      maxLength: 255
      pattern: "^[!-~]+$"
    ExternalId:
      type: string
      minLength: 1
      maxLength: 255
      pattern: "^[!-~]+$"
...
    ActorRef:
      type: object
      additionalProperties: false
      required:
      - type
      - id
      properties:
        type:
          type: string
          enum:
          - carbon
          - silicon
          - application
          - service
        id:
          "$ref": "#/components/schemas/ExternalId"
    HookStatus:
      type: string
      description: Disabled is reversible and retained; deleted is the separate 45-day
        recovery lifecycle.
      enum:
      - active
      - disabled
      - deleted
    SignatureAlgorithm:
      type: string
      enum:
      - HMAC-SHA1
      - HMAC-SHA256
      - HMAC-SHA384
      - HMAC-SHA512
      - SHA1
      - SHA256
      - SHA384
      - SHA512
      - Ed25519
      - ECDSA-SHA256
      - RSA-SHA1
      - RSA-SHA256
    SignatureEncoding:
      type: string
      enum:
      - hex
      - base64
      - base64url
      - raw
    SecretEncoding:
      type: string
      enum:
      - utf8
      - ascii
      - hex
      - base64
      - base64url
      - raw
    SignatureExpression:
      type: string
      minLength: 1
      maxLength: 4096
      description: An expression over request blocks (`request.raw_body`, `request.headers["name"]`,
        `request.query.key`, `request.body.path`, `request.form.key`, `request.multipart.key`,
        `request.method`, `request.url`, ...), `hook.id`, `hook.url`, `secret`, and
        `key.public`, combined with concat, join, sort, sort_keys, utf8, ascii, url_encode,
        url_decode, percent_encode, percent_decode, canonicalize_url, canonicalize_query,
        json_encode, form_encode, sha1, sha256, sha384, sha512, hex, hex_decode, base64,
        base64_decode, base64url, base64url_decode, lowercase, uppercase, and trim.
    SignaturePolicy:
      type: object
      additionalProperties: false
      required:
      - required
      - algorithm
      - payload
      - signature
      - signature_encoding
      - secret_encoding
      - public_key
      - has_secret
      properties:
        required:
          type: boolean
        algorithm:
          "$ref": "#/components/schemas/SignatureAlgorithm"
        payload:
          "$ref": "#/components/schemas/SignatureExpression"
        signature:
          "$ref": "#/components/schemas/SignatureExpression"
        signature_encoding:
          "$ref": "#/components/schemas/SignatureEncoding"
        secret_encoding:
          "$ref": "#/components/schemas/SecretEncoding"
        public_key:
          type:
          - string
          - 'null'
          maxLength: 16384
        has_secret:
          type: boolean
    SignaturePolicyInput:
      type: object
      additionalProperties: false
      description: 'Every member is optional. Defaults are the Standard Webhooks convention:
        required, HMAC-SHA256, concat(request.headers["webhook-id"], ".", request.headers["webhook-timestamp"],
        ".", request.raw_body), request.headers["webhook-signature"], base64, utf8.
        BYOS secrets are stored verbatim on creation and replace the current secret on
        update. Omitting secret generates one on creation and keeps the current
        secret on update. Generated and rotated secrets respect secret_encoding.'
      properties:
        required:
          type: boolean
        algorithm:
          "$ref": "#/components/schemas/SignatureAlgorithm"
        payload:
          "$ref": "#/components/schemas/SignatureExpression"
        signature:
          "$ref": "#/components/schemas/SignatureExpression"
        signature_encoding:
          "$ref": "#/components/schemas/SignatureEncoding"
        secret_encoding:
          "$ref": "#/components/schemas/SecretEncoding"
        public_key:
          type:
          - string
          - 'null'
          maxLength: 16384
        secret:
          type: string
          minLength: 1
          maxLength: 4096
          writeOnly: true
          description: BYOS secret text (1-4096 UTF-8 bytes, no control characters).
            Set at creation or replace after registration. Must decode to nonempty
            key bytes with secret_encoding. Not accepted for asymmetric algorithms.
            Never echoed after the creation response.
    HookBase:
      type: object
      required:
      - id
      - org_id
      - silicon_id
      - name
      - endpoint_url
      - endpoint_key
      - status
      - signature
      - time_zone
      - created_by
      - created_at
      properties:
        id:
          type: string
          format: uuid
        org_id:
          "$ref": "#/components/schemas/OrganizationId"
        silicon_id:
          "$ref": "#/components/schemas/SiliconId"
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type:
          - string
          - 'null'
          maxLength: 2000
        endpoint_url:
          type: string
          format: uri
        endpoint_key:
          type: string
          pattern: "^[A-Z0-9]{8}$"
        status:
          "$ref": "#/components/schemas/HookStatus"
        signature:
          "$ref": "#/components/schemas/SignaturePolicy"
        time_zone:
          type: string
          maxLength: 64
          description: IANA zone used in delivery summaries.
        created_by:
          "$ref": "#/components/schemas/ActorRef"
        created_at:
          type: string
          format: date-time
        disabled_at:
          type:
          - string
          - 'null'
          format: date-time
        deleted_at:
          type:
          - string
          - 'null'
          format: date-time
        recoverable_until:
          type:
          - string
          - 'null'
          format: date-time
        last_received_at:
          type:
          - string
          - 'null'
          format: date-time
          description: When the provider last reached out with a verified request.
        last_blocked_at:
          type:
          - string
          - 'null'
          format: date-time
        endpoint_rotated_at:
          type:
          - string
          - 'null'
          format: date-time
    Hook:
      type: object
      allOf:
      - "$ref": "#/components/schemas/HookBase"
      unevaluatedProperties: false
    HookWithSecret:
      type: object
      allOf:
      - "$ref": "#/components/schemas/HookBase"
      required:
      - signing_secret
      properties:
        signing_secret:
          type:
          - string
          - 'null'
          description: Generated secrets are `v1.` followed by 32 alphanumeric characters;
            a supplied secret is echoed as given. Null for asymmetric algorithms.
      unevaluatedProperties: false
    HookCreate:
      type: object
      additionalProperties: false
      required:
      - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type:
          - string
          - 'null'
          maxLength: 2000
        time_zone:
          type: string
          maxLength: 64
          default: UTC
        signature:
          "$ref": "#/components/schemas/SignaturePolicyInput"
...
    HooksEnabledUpdate:
      type: object
      additionalProperties: false
      required:
      - hook_ids
      - enabled
      properties:
        hook_ids:
          type: array
          minItems: 1
          maxItems: 1000
          uniqueItems: true
          items:
            type: string
            format: uuid
        enabled:
          type: boolean
    HookPage:
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          maxItems: 1000
          items:
            "$ref": "#/components/schemas/Hook"
...
    Error:
      type: object
      additionalProperties: false
      required:
      - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
          - code
          - message
          - request_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type: string
    TestEnvironment:
      type: object
      required:
      - id
      - org_id
      - creator_kind
      - creator_id
      - name
      - generation
      - created_at
      - last_activity_at
      - deleted_at
      properties:
        id:
          type: string
          format: uuid
        org_id:
          "$ref": "#/components/schemas/OrganizationId"
        creator_kind:
          type: string
          enum:
          - carbon
          - silicon
        creator_id:
          type: string
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type:
          - string
          - 'null'
          maxLength: 2000
        generation:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        last_activity_at:
          type: string
          format: date-time
        deleted_at:
          type:
          - string
          - 'null'
          format: date-time
...